chore(vscode): Use File Watcher extension
While developing, save Dockerfiles or entrypoint shell scripts
are very time consuming, we need to save and run command
manually on your terminal. With File Watcher, I just needed to config
some things to be able to save a file and run task
scripts automatically.

It was a game changer in developing Docker images.