fix(nginx): Fix user running entrypoint
Default from fpm-* is to run entrypoint as user 1000, to avoid
permission issues while developing and installing
new packages, for example.

But nginx command needs to be run as root. I'll need a fix after this
commit to make this change specifically for nginx and supervisor,
while other commands must be run and will make the permission
error returns.