feat: Use php-fpm as CMD for quality
It allows us to reuse quality tag in many jobs of a Gitlab CI pipeline.
So we don't need to pull fpm-dev and quality tags, as both does the same
and quality has additional tools.

Now I'm able to keep a running php-fpm server along with doing
php-cs-fixer or phpstan analysis.